+++ Editor/References/RuntimeReferenceDrawer.cs
@@ -4,11 +4,16 @@ using UnityEngine;
 namespace UnderLogic.Variables.References.Editor
 {
     [CustomPropertyDrawer(typeof(BoolReference), true)]
+    [CustomPropertyDrawer(typeof(ColorReference), true)]
     [CustomPropertyDrawer(typeof(DoubleReference), true)]
     [CustomPropertyDrawer(typeof(FloatReference), true)]
     [CustomPropertyDrawer(typeof(IntReference), true)]
     [CustomPropertyDrawer(typeof(StringReference), true)]
     [CustomPropertyDrawer(typeof(TransformReference), true)]
+    [CustomPropertyDrawer(typeof(Vector2IntReference), true)]
+    [CustomPropertyDrawer(typeof(Vector2Reference), true)]
+    [CustomPropertyDrawer(typeof(Vector3Reference), true)]
+    [CustomPropertyDrawer(typeof(Vector4Reference), true)]
     public class RuntimeReferenceDrawer : PropertyDrawer
     {
         private SerializedProperty _useConstant;