+++ Runtime/References/Vector2IntReference.cs
@@ -1,0 +1,8 @@
+using System;
+using UnityEngine;
+
+namespace UnderLogic.Variables.References
+{
+    [Serializable]
+    public class Vector2IntReference : RuntimeReference<Vector2Int, Vector2IntVariable> { }
+}