+++ Runtime/References/Vector2Reference.cs
@@ -1,0 +1,8 @@
+using System;
+using UnityEngine;
+
+namespace UnderLogic.Variables.References
+{
+    [Serializable]
+    public class Vector2Reference : RuntimeReference<Vector2, Vector2Variable> { }
+}