+++ Runtime/References/Vector3Reference.cs
@@ -1,0 +1,8 @@
+using System;
+using UnityEngine;
+
+namespace UnderLogic.Variables.References
+{
+    [Serializable]
+    public class Vector3Reference : RuntimeReference<Vector3, Vector3Variable> { }
+}