+++ Runtime/References/Vector4Reference.cs
@@ -1,0 +1,8 @@
+using System;
+using UnityEngine;
+
+namespace UnderLogic.Variables.References
+{
+    [Serializable]
+    public class Vector4Reference : RuntimeReference<Vector4, Vector4Variable> { }
+}