+++ Runtime/Watchers/Vector2IntVariableWatcher.cs
@@ -1,0 +1,6 @@
+using UnityEngine;
+
+namespace UnderLogic.Variables.Watchers
+{
+    public class Vector2IntVariableWatcher : RuntimeVariableWatcher<Vector2Int, Vector2IntVariable> { }
+}