+++ Runtime/Watchers/Vector2VariableWatcher.cs
@@ -1,0 +1,6 @@
+using UnityEngine;
+
+namespace UnderLogic.Variables.Watchers
+{
+    public class Vector2VariableWatcher : RuntimeVariableWatcher<Vector2, Vector2Variable> { }
+}